+++ src/utils/get.ts
@@ -1,6 +1,6 @@
 import isUndefined from './isUndefined';
 import isNullOrUndefined from './isNullOrUndefined';
-import filterOutFalsy from './filterOutFalsy';
+import filterOutFalsy from './uniq';
 
 export default (obj: any, path: string, defaultValue?: unknown) => {
   const result = filterOutFalsy(path.split(/[,[\].]+?/)).reduce(
