+++ src/utils/remove.ts
@@ -1,6 +1,6 @@
 import isUndefined from './isUndefined';
 import isArray from './isArray';
-import filterOutFalsy from './filterOutFalsy';
+import uniq from './uniq';
 
 const removeAt = <T>(data: T[], index: number): T[] => [
   ...data.slice(0, index),
@@ -16,7 +16,7 @@ function removeAtIndexes<T>(data: T[], index: number[]): T[] {
     }
   }
 
-  return filterOutFalsy(data);
+  return uniq(data);
 }
 
 export default <T>(data: T[], index?: number | number[]): T[] =>
